requirements.txt
aiohttp==3.6.2
ansible==2.8.8
appdirs==1.4.3
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
distlib==0.3.0
filelock==3.0.12
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
Jinja2==2.11.2
joblib==0.15.1
lazy-object-proxy==1.4.3
livereload==2.6.1
lunr==0.5.8
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mercurial==5.3.2
mkdocs==1.1.2
mpmath==1.1.0
multidict==4.7.6
netaddr==0.7.19
nltk==3.5
packaging==20.4
pbr==5.4.5
pkgconfig==1.5.1
pnglatex==1.1
powerline-status===2.8.1.dev9999-git.b-c26938a9809bf8b55399bedec6f4e6300650f5ed-
psutil==5.7.0
pycparser==2.20
pylint==2.5.2
pyparsing==2.4.7
pytexit==0.3.3
PyYAML==5.3.1
regex==2020.5.14
six==1.15.0
slackclient==2.6.2
stevedore==1.32.0
sympy==1.6
tex2pix==0.3.1
toml==0.10.1
tornado==6.0.4
tqdm==4.46.0
typed-ast==1.4.1
virtualenv==20.0.20
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualfish==2.2.3
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
